+++ src/MySocailApp.Domain/BalanceAggregate/Abstracts/IBalanceRepository.cs
@@ -8,6 +8,6 @@ namespace MySocailApp.Domain.BalanceAggregate.Abstracts
         Task<Balance> GetAsync(int id, CancellationToken cancellationToken);
         Task<bool> HasBalance(int id, CancellationToken cancellationToken);
         Task<bool> HasDept(int id, CancellationToken cancellationToken);
-        Task<decimal> GetBalanceAsync(int id, CancellationToken cancellationToken);
+        Task<int> GetBalanceAsync(int id, CancellationToken cancellationToken);
     }
 }
